requirements.txt
discord.py==2.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.61.0
googletrans==4.0.0rc1
grpcio==1.59.0
grpcio-status==1.59.0
PyNaCl==1.5.0
wavelink==3.2.0
wikipedia==1.4.0